x_network/requirements.txt
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2
django-crispy-forms==1.9.0
django-formtools==2.2
django-import-export==2.0.2
django-rest-framework==0.1.0
django-reversion==3.0.7
djangorestframework==3.11.0
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.9.2
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
psycopg2==2.8.4
psycopg2-binary==2.8.4
pytz==2019.3
PyYAML==5.3
sqlparse==0.3.1
tablib==1.1.0
xadmin2==2.0.3
xlrd==1.2.0
xlwt==1.3.0
